+++ pkg/server/exec/errors.go
@@ -8,4 +8,5 @@ import (
 var (
 	ErrDeployNotFound  = status.Errorf(codes.NotFound, "Current deploy not found")
 	ErrNonZeroExitCode = status.Errorf(codes.Unknown, "Exec command returned a non zero value")
+	ErrTimeout         = status.Errorf(codes.Aborted, "Timeout on performing the command")
 )