+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [NEXT_RELEASE]
+### Fixed
+- Return error with timeout message when there's no resources for build, release phase and command exec
+
 ## [0.18.0] - 2018-03-27
 ### Added
 - `service` command, for now only used to enable ssl support